+++ src/main/java/com/suitecare/suitecare/api/member/dto/ChangePasswordRequestDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ChangePasswordRequestDTO {
-    private int login_id;
+    private Long login_id;
     private String originPassword;
     private String newPassword;
     private String newPasswordCheck;
